refactor: 로직 분리 및 누락 아이콘 추가 (#45)
- 비즈니스 로직 분리
- 누락된 아이콘 추가
