+++ src/constants/my-page/gauge.ts
@@ -1,0 +1,10 @@
+import { TasteScore } from '@/types/gauge-bar/tasteTypes'
+
+export const GAUGE_VALUES: TasteScore[] = [
+  { type: 'sweetness_level', score: 4 },
+  { type: 'acidity_level', score: 3.5 },
+  { type: 'body_level', score: 5 },
+  { type: 'carbonation_level', score: 1.5 },
+  { type: 'bitterness_level', score: 2 },
+  { type: 'aroma_level', score: 3.5 },
+]
